+++ lib/pbio/platform/city_hub/pbdrvconfig.h
@@ -21,6 +21,7 @@
 #define PBDRV_CONFIG_BUTTON                         (1)
 #define PBDRV_CONFIG_BUTTON_GPIO                    (1)
 #define PBDRV_CONFIG_BUTTON_GPIO_NUM_BUTTON         (1)
+#define PBDRV_CONFIG_BUTTON_GPIO_DEBOUNCE           (1)
 
 #define PBDRV_CONFIG_BLOCK_DEVICE                   (1)
 #define PBDRV_CONFIG_BLOCK_DEVICE_FLASH_STM32       (1)