+++ lib/pbio/platform/essential_hub/pbdrvconfig.h
@@ -45,6 +45,7 @@
 #define PBDRV_CONFIG_BUTTON                         (1)
 #define PBDRV_CONFIG_BUTTON_GPIO                    (1)
 #define PBDRV_CONFIG_BUTTON_GPIO_NUM_BUTTON         (1)
+#define PBDRV_CONFIG_BUTTON_GPIO_DEBOUNCE           (1)
 
 #define PBDRV_CONFIG_CHARGER                        (1)
 #define PBDRV_CONFIG_CHARGER_MP2639A                (1)