+++ lib/pbio/platform/technic_hub/pbdrvconfig.h
@@ -34,6 +34,7 @@
 #define PBDRV_CONFIG_BUTTON                         (1)
 #define PBDRV_CONFIG_BUTTON_GPIO                    (1)
 #define PBDRV_CONFIG_BUTTON_GPIO_NUM_BUTTON         (1)
+#define PBDRV_CONFIG_BUTTON_GPIO_DEBOUNCE           (1)
 
 #define PBDRV_CONFIG_CLOCK                          (1)
 #define PBDRV_CONFIG_CLOCK_STM32                    (1)